refactor: DateFormatter static으로 구현 (#34)
